feat(对话): 修复消息未保存问题，将保存消息提取MessageService进行

+++ ai-chat-modules/ai-chat-chat/src/main/java/cn/com/chat/chat/service/IChatMessageService.java
@@ -61,6 +61,6 @@ public interface IChatMessageService {
 
     ChatMessageVo queryByMessageId(String messageId);
 
-    void updateStatusByMessageId(String messageId, Integer status);
+    void updateStatusByMessageId(String messageId, Long status);
 
 }
